requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.0.3
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.5.0
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
